+++ HAL/brake.h
@@ -6,7 +6,7 @@
 
 void g_config_BRAKE(void);
 void g_update_brake_state(void);
-u8 g_get_brake_state(void);
+u32 g_get_brake_state(void);
 
 #endif
